SmartHome
1. When the system is started, "Smart Home" will be displayed on the LCD.
2. When the system starts, a timer will start and will be displayed on the 7-segment display
("HH.MM"), where every minute will be equal to one second of reality. When the timer
arrives at 23:59, will be reset.
3. The system will have 2 states: "closed" / "open". As long as it is open, an LED will light up
green and the system will work normally. In the "closed" state, a red LED and all will be lit.
the functionalities will be stopped (setting the alarm and displaying / changing the temperature), with
except for the stopwatch.
4. The system will be "closed" between 22: 00-06: 00 and "open" in the rest of the interval.
5. The current status will be displayed on row 2 of the LCD display.
6. The temperature adjustment will be simulated with the help of the potentiometer (as in the laboratories
before). A threshold of 27 degrees will be set. When this threshold is exceeded, it is
an alarm (buzzer) will start and, optionally, a DC motor will start. The temperature will be displayed on
first row of the LCD ("T = gg.g (ASCII) degrees C")
7. Pressing a button will enter the alarm setting mode. The time will come
the active alarm will be set with a button / potentiometer. When the timer arrives
at the set time, even if the system is closed, the first line of the LCD will be displayed
"ALARM!" for 5 seconds and then return to the previous state.